Project Maintence
This commit combined with previous commit to this branch fixes the Launcher project not loading.

ClassLibrary1 project still removed along with the .NET Framework version of the launcher.

Reference to the GUI project from the .NETCore project removed to allow build in current state. This will need to be added in at a later date.